requirements.txt
PySide6==6.11.0
SQLAlchemy==2.0.49
alembic==1.18.4
psycopg[binary]==3.3.3
python-dotenv==1.2.2
openpyxl==3.1.5
defusedxml==0.7.1
reportlab==4.4.10
pytest==9.0.3
pytest-qt==4.5.0
bcrypt==5.0.0
pyinstaller==6.19.0